requirements.txt

eth-brownie>=1.19.0
web3>=6.0.0
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
asyncio
sqlite3
